Fix bug occurring when keeping a key down
When keeping [Space] or [DownArrow] down, it might happen
that the falling tetrimino has not yet been instanciated after being
dropped, causing an error when the following event handler tries to
perform an action on it.
This commit basically just adds a check that it exists before doing
anything with it.